backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.0.0
tortoise-orm==0.20.0
aiomysql==0.2.0
redis==5.0.0
httpx==0.24.1
bcrypt==4.0.1
python-dotenv==1.0.0
pytest==7.4.2
pytest-asyncio==0.21.1
loguru==0.7.0 
openpyxl
xlrd
pandas
xlsxwriter
# 数据生成所需依赖
numpy
faker